Create README.md
The package provides simple commands that add functionality to the standard classes which are available with KOMA-classes for a long time. This eases the possible switch to the great KOMA-classes.

For example, `addchap` adds an unnumbered chapter *and* creates an entry in the table of contents as well as sets the marks (to be used in headers) accordingly. The starred form of `addchap` works like the classical `chapter*` but clears the mark. That prevents you from having the title of your last chapter in the header. 

More information can be found in the documentation of KOMA-script.

This package is by far not complete. A few features will be added in the future to further ease the change. But to be honest, if you want to have the functionality of KOMA-script, use KOMA-script ;-)